fix: massive restructuring of code modules
