+++ README.md
@@ -1,2 +1,4 @@
 # 5k
-This is a dicegame where players take turns to roll dice, score points and the first player to 5k wins the game
+This is a dicegame where players take turns to roll dice, score points and the first player to 5k wins the game.
+
+I have halted this project for now. Looking to make progress and restructure.
