+++ src/welcome/welcome_banner.txt
